refactor: rename all example classes and functions using a "demo" prefix.
By that way, it will be easier to follow them and to remove them when they will become useless.
