+++ bindings/R/testBinding/R/manualWrapper.R
@@ -1,10 +1,10 @@
 # wrapper function to invoke test_binding
-test_binding1 <- function() {
-  result <- .Call("test_binding1")
+demo_binding1 <- function() {
+  result <- .Call("demo_binding1")
   return(result)
 }
 
-test_binding2 <- function() {
-  result <- .Call("test_binding2")
+demo_binding2 <- function() {
+  result <- .Call("demo_binding2")
   return(result)
 }
