+++ bindings/R/testBinding/src/demo_manual_binding2.cpp
@@ -2,6 +2,9 @@
  R callable function
  ***************************************************************************/
 
+// Required since libKriging.hpp has signatures with armadillo types
+#include <RcppArmadillo.h>
+
 #include <R.h>
 #include <Rdefines.h>
 
@@ -10,21 +13,20 @@
 
 #include <memory>
 
-#include <libKriging/TestClass.hpp>
+#include <libKriging/libKriging.hpp>
 
-extern "C" SEXP test_binding1() {
+extern "C" SEXP demo_binding2() {
 
     //    std::cout << "libKriging class tests" << std::endl;
     printf("libKriging class tests\n");
 
-    std::unique_ptr<TestClass> x(new TestClass());
-    int tmp = [&x]() { return x->f(); }();
+    std::unique_ptr<DemoArmadilloClass> x(new DemoArmadilloClass());
 
     SEXP result;
 
     PROTECT(result = NEW_INTEGER(2));
     INTEGER(result)[0] = (int) 1;
-    INTEGER(result)[1] = (int) tmp;
+    INTEGER(result)[1] = (int) 2;
     UNPROTECT(1);
 
     return (result);
