+++ bindings/R/testBinding/src/demo_rcpp_binding.cpp
@@ -1,0 +1,7 @@
+#include <RcppArmadillo.h>
+#include "libKriging/DemoArmadilloClass.hpp"
+
+// [[Rcpp::export]]
+arma::vec getEigenValues(arma::mat M) {
+  return DemoArmadilloClass::getEigenValues(M);
+}
