--- bindings/R/testBinding/src/test_rcpp_binding.cpp
@@ -1,7 +1,0 @@
-#include <RcppArmadillo.h>
-#include "libKriging/ArmadilloTestClass.hpp"
-
-// [[Rcpp::export]]
-arma::vec getEigenValues(arma::mat M) {
-  return ArmadilloTestClass::getEigenValues(M);
-}
