+++ bindings/R/testBinding/test/test.R
@@ -1,9 +1,9 @@
 library(testBinding)
-test_binding1()
-test_binding2()
+demo_binding1()
+demo_binding2()
 
 
- set.seed(42)
+set.seed(42)
 X <- matrix(rnorm(4*4), 4, 4)
 Z <- X %*% t(X)
 getEigenValues(Z)
