+++ src/lib/DemoFunction.cpp
@@ -1,4 +1,4 @@
-#include "libKriging/TestFunction.hpp"
+#include "libKriging/DemoFunction.hpp"
 
 LIBKRIGING_EXPORT int f() {
   return 42;
