--- src/lib/include/libKriging/TestFunction.hpp
@@ -1,8 +1,0 @@
-#ifndef LIBKRIGING_TESTFUNCTION_HPP
-#define LIBKRIGING_TESTFUNCTION_HPP
-
-#include "libKriging_exports.h"
-
-LIBKRIGING_EXPORT int f();
-
-#endif  // LIBKRIGING_TESTFUNCTION_HPP
