+++ tests/armadillo_example.cpp
@@ -1,6 +1,6 @@
-#include "libKriging/ArmadilloTestClass.hpp"
 #include <memory>
+#include "libKriging/DemoArmadilloClass.hpp"
 
 int main() {
-  std::unique_ptr<ArmadilloTestClass> x(new ArmadilloTestClass());
+  std::unique_ptr<DemoArmadilloClass> x(new DemoArmadilloClass());
 }
