+++ tests/class_unit_test.cpp
@@ -1,13 +1,13 @@
 #include <iostream>
 
-#include "libKriging/TestClass.hpp"
+#include "libKriging/DemoClass.hpp"
 
 #include <memory>
 
 int main() {
   std::cout << "libKriging class tests" << std::endl;
 
-  std::unique_ptr<TestClass> x(new TestClass());
+  std::unique_ptr<DemoClass> x(new DemoClass());
   std::cout << x->name() << std::endl;
 
   return 0;
